@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0  <3.0.0"
+  sdk: ">=2.14.0  <3.0.0"
 
 dependencies:
   cupertino_icons: ^1.0.2
