@@ -3,7 +3,7 @@ description: A new Flutter package project.
 version: 0.0.1
 # homepage:
 environment:
-  sdk: ">=2.12.0  <3.0.0"
+  sdk: ">=2.14.0  <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
