@@ -5,12 +5,17 @@ class Source {
   String? name;
   String? category;
   String? country;
-
+  String? description;
+  String? url;
+  String? language;
   Source({
     this.id,
     this.name,
     this.category,
     this.country,
+    this.description,
+    this.url,
+    this.language,
   });
 
   Source copyWith({
@@ -18,12 +23,18 @@ class Source {
     String? name,
     String? category,
     String? country,
+    String? description,
+    String? url,
+    String? language,
   }) {
     return Source(
       id: id ?? this.id,
       name: name ?? this.name,
       category: category ?? this.category,
       country: country ?? this.country,
+      description: description ?? this.description,
+      url: url ?? this.url,
+      language: language ?? this.language,
     );
   }
 
@@ -33,6 +44,9 @@ class Source {
       'name': name,
       'category': category,
       'country': country,
+      'description': description,
+      'url': url,
+      'language': language,
     };
   }
 
@@ -42,6 +56,9 @@ class Source {
       name: map['name'],
       category: map['category'],
       country: map['country'],
+      description: map['description'],
+      url: map['url'],
+      language: map['language'],
     );
   }
 
@@ -51,7 +68,7 @@ class Source {
 
   @override
   String toString() {
-    return 'Source(id: $id, name: $name, category: $category, country: $country)';
+    return 'Source(id: $id, name: $name, category: $category, country: $country, description: $description, url: $url, language: $language)';
   }
 
   @override
@@ -62,11 +79,20 @@ class Source {
         other.id == id &&
         other.name == name &&
         other.category == category &&
-        other.country == country;
+        other.country == country &&
+        other.description == description &&
+        other.url == url &&
+        other.language == language;
   }
 
   @override
   int get hashCode {
-    return id.hashCode ^ name.hashCode ^ category.hashCode ^ country.hashCode;
+    return id.hashCode ^
+        name.hashCode ^
+        category.hashCode ^
+        country.hashCode ^
+        description.hashCode ^
+        url.hashCode ^
+        language.hashCode;
   }
 }
